Move redux component to redux folder and fix files paths
