+++ app/javascript/application.js
@@ -3,7 +3,7 @@ import "./controllers";
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App';
-import store from './store';
+import store from './redux/store';
 import { Provider } from 'react-redux';
 
 ReactDOM.render(
