+++ app/javascript/redux/actions/types.js
