+++ app/javascript/redux/reducers/index.js
